FADING WORK DE LA MORT DE TOM QUI TUE ET MEME MAXIME AUSSI

+++ 4A3DJVCompoImage/AlgoImages.h
@@ -19,7 +19,7 @@ public:
 	static void getBackground(std::vector<Image> images, Image& res);
 	static bool checkSizeImages(std::vector<Image> images);
 	static void getImageMask(Image targetImage, Image background, Image& mask, float maxDiff);
-	static void binaryMerge(Image* mask, Image* image1, Image* image2, int alpha);
+	static void binaryMerge(Image* mask, Image* image1, Image* image2, float alpha);
 	static void getVideoFrame(std::string outputPath, std::string filename, int step);
 	static bool checkVideoInPath(std::string path, std::string videoName);
 	static std::vector<Image> selectUsedImages(std::vector<Image>& images, Settings* settings);
